apps/mcp-server/test/requirements.txt
aperturedb
pytest
pytest-asyncio
fastmcp
pillow
numpy

